+++ SourceCode/Math/Easing.h
@@ -2,8 +2,8 @@
 #include <math.h>
 #define PI 3.14f
 
-double easeInSine(double start, double end, double movetime, double endtime);
-double easeInQuint(double start, double end, double movetime, double endtime);
-double easeOutSine(double start, double end, double movetime, double endtime);
-double easeOutQuint(double start, double end, double movetime, double endtime);
+float easeInSine(float start, float end, float movetime, float endtime);
+float easeInQuint(float start, float end, float movetime, float endtime);
+float easeOutSine(float start, float end, float movetime, float endtime);
+float easeOutQuint(float start, float end, float movetime, float endtime);
 double EaseInBack(float t, int start, int end, int num = 1, double p1 = 1.70158);
